pyproject.toml
[project]
name = "medication-augmentation"
version = "1.0.0"
description = "Intelligent medication discovery and classification system for clinical research"
readme = "README.md"
authors = [
    { name = "Patrick Sheehan", email = "pat@unlearn.ai" }
]
requires-python = ">=3.11"
dependencies = [
    # CLI Framework
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    # Data Processing
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "pydantic>=2.5.0",
    "pyreadstat>=1.3.0",  # SAS file support
    # Web Scraping & HTTP
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    # LLM Integration
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    # Caching & Storage
    "diskcache>=5.6.0",
    # Configuration & Utilities
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
]

[project.scripts]
med-aug = "med_aug.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/med_aug"]

[dependency-groups]
dev = [
    "pandas>=2.3.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
